chore: Update dependencies and refactor login page
