+++ tsconfig.json
@@ -5,6 +5,8 @@
 		"lib": ["ES2020", "DOM", "DOM.Iterable"],
 		"module": "ESNext",
 		"skipLibCheck": true,
+		"allowSyntheticDefaultImports": true,
+		"esModuleInterop": true,
 
 		/* Bundler mode */
 		"moduleResolution": "node",
